updated structure of dto directory
