+++ backend/src/main/java/com/cryptowatch/backend/controller/AdminController.java
@@ -1,7 +1,7 @@
 package com.cryptowatch.backend.controller;
 
-import com.cryptowatch.backend.dto.ExportDataResponse;
-import com.cryptowatch.backend.dto.ImportDataResponse;
+import com.cryptowatch.backend.dto.response.ExportDataResponse;
+import com.cryptowatch.backend.dto.response.ImportDataResponse;
 import com.cryptowatch.backend.service.AdminService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
