+++ backend/src/main/java/com/cryptowatch/backend/dto/cmc/CmcHistoricalResponse.java
@@ -1,0 +1,7 @@
+package com.cryptowatch.backend.dto.cmc;
+
+import lombok.Data;
+@Data
+public class CmcHistoricalResponse {
+    private CmcHistoricalData data;
+}
