+++ backend/src/main/java/com/cryptowatch/backend/dto/cmc/CmcInfo.java
@@ -1,0 +1,10 @@
+package com.cryptowatch.backend.dto.cmc;
+
+import lombok.Data;
+
+@Data
+public class CmcInfo {
+    private int id;
+    private String name;
+    private String symbol;
+}
