+++ backend/src/main/java/com/cryptowatch/backend/dto/cmc/CmcQuoteData.java
@@ -1,0 +1,9 @@
+package com.cryptowatch.backend.dto.cmc;
+
+import lombok.Data;
+import java.util.Map;
+
+@Data
+public class CmcQuoteData {
+    private Map<String, CmcUsd> quote;
+}
