+++ backend/src/main/java/com/cryptowatch/backend/dto/cmc/CmcResponse.java
@@ -1,0 +1,9 @@
+package com.cryptowatch.backend.dto.cmc;
+
+import lombok.Data;
+import java.util.Map;
+
+@Data
+public class CmcResponse {
+    private Map<String, CmcQuoteData> data;
+}
