+++ backend/src/main/java/com/cryptowatch/backend/dto/common/PresetDto.java
@@ -1,4 +1,4 @@
-package com.cryptowatch.backend.dto;
+package com.cryptowatch.backend.dto.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
