+++ backend/src/main/java/com/cryptowatch/backend/dto/request/SavePresetRequest.java
@@ -1,4 +1,4 @@
-package com.cryptowatch.backend.dto;
+package com.cryptowatch.backend.dto.request;
 
 import lombok.Data;
 import java.util.Date;
