+++ backend/src/main/java/com/cryptowatch/backend/dto/response/DeleteCoinResponse.java
@@ -1,0 +1,15 @@
+package com.cryptowatch.backend.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class DeleteCoinResponse {
+    private boolean success;
+    private String message;
+}
