+++ backend/src/main/java/com/cryptowatch/backend/dto/response/ExportDataResponse.java
@@ -1,19 +1,20 @@
-package com.cryptowatch.backend.dto;
+package com.cryptowatch.backend.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.util.Date;
+import java.util.List;
 import java.util.Map;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ImportDataResponse {
+public class ExportDataResponse {
     private boolean success;
-    private String message;
+    private Map<String, List<?>> data; // keys: users, coins_meta, coin_snapshots, statistics_settings
+    private Date exportedAt;
     private Map<String, Long> recordCount;
-    private Date importedAt;
 }
