+++ backend/src/main/java/com/cryptowatch/backend/dto/response/FavoritesResponse.java
@@ -1,4 +1,4 @@
-package com.cryptowatch.backend.dto;
+package com.cryptowatch.backend.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,13 +6,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.util.List;
 
+import com.cryptowatch.backend.dto.common.CoinDto;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PresetListResponse {
+public class FavoritesResponse {
     private boolean success;
-    private List<PresetDto> presets;
+    private List<CoinDto> coins;
     private long totalCount;
     private int pageSize;
     private int pageNo;
