+++ backend/src/main/java/com/cryptowatch/backend/dto/response/SearchCoinsResponse.java
@@ -1,4 +1,4 @@
-package com.cryptowatch.backend.dto;
+package com.cryptowatch.backend.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,6 +6,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.util.List;
 import java.util.Map;
+
+import com.cryptowatch.backend.dto.common.CoinDto;
 
 @Data
 @Builder
